Перенесли функционал контроллера и отрефакторили код 5.1
